Movement prediction
Calculating a circle based on character movement. Predicting future positions with angular velocity. Drawing debug circles.

For whatever reason I couldn't find a good solution online in a reasonable time, so I decided to just invent how to do it myself. It's been a while since I did geometry and this was somewhat different as to what I did in school.