Wombat Highlight Button
Clicking the enemy button will highlight the enemy, and will change the button style to show which enemies are currently highlighted. Closing the window will unhighlight enemies, as will pressing the "Get Spawned Enemies" button.
For whatever reason, when trying to change the button style it will crash the editor, but only if you try to use a custom style.